chore: separate types according to usage
