+++ resources/js/types/index.d.ts
@@ -3,6 +3,7 @@ import { User } from "@type/User";
 export type PageProps<
     T extends Record<string, unknown> = Record<string, unknown>
 > = T & {
+    appName: string;
     auth: {
         user: User;
     };
