+++ resources/js/types/model/Children.d.ts
