+++ resources/js/types/model/Schedule.d.ts
