+++ resources/js/types/model/UserCourse.d.ts
